Basic Synonym Recognition
It's working baby. Stored resource files inside the .exe. My only
thought it that it seems like visual studio is including a lot of crap
in my commits, which is annoying. Lots of reworking on project
internals. You can type commands to see a basic output of if they're in
the game yet, go ahead, try. Stuff like "chop tree" a return value of -1
means it's not in.