Added hunger, health, world saving & loading
Commands "save" and "load" are now a thing. Not sure how well
implemented it is, but whatevah. Hunger and health bars are in as well.
Both are set on a scale of 1000, and every action will require varying
amounts of hunger. Currently hunger can only be regained by eating berry
bushes. Blocks are now registered a little more fancily as well.