Major game components complete
The game is pretty much done. Balance, bugfixing fleshing out, and
playtesting is still needed. Also added a python script to generate the
code for a boat in case I ever want to change it. There are a few places
in this commit that the implementation methods are not of high standard.